// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  PRODUCT_LIKED = 'product_liked',
  PRODUCT_COMMENTED = 'product_commented',
  REVIEW_RECEIVED = 'review_received',
  NEW_FOLLOWER = 'new_follower',
  ORDER_CREATED = 'order_created',
  ORDER_CONFIRMED = 'order_confirmed',
  ORDER_SHIPPED = 'order_shipped',
  ORDER_DELIVERED = 'order_delivered',
  ORDER_COMPLETED = 'order_completed',
  ORDER_CANCELLED = 'order_cancelled',
  REFUND_COMPLETED = 'refund_completed',
  REWARD_CREDITED = 'reward_credited',
  REWARD_NOT_DETECTED = 'reward_not_detected',
  REWARD_PROCESSING_FAILED = 'reward_processing_failed',
  POINTS_REFUNDED = 'points_refunded',
  SALE_POINTS_RELEASED = 'sale_points_released',
  NEW_MESSAGE = 'new_message',
  SYSTEM_ANNOUNCEMENT = 'system_announcement',
  ACCOUNT_STATUS_CHANGED = 'account_status_changed',
  SECURITY_ALERT = 'security_alert',
  SELLER_APPLICATION_APPROVED = 'seller_application_approved',
  SELLER_APPLICATION_REJECTED = 'seller_application_rejected',
}